fix(ada_boost, weak_learner): improved weights and beta calculations in order to increase the loss function
The idea is that having a very high cumulative loss (~1*dataset.length) as well as a big number of samples caused (1) the loss to be too small to be effective, because of very small weight initialization, and two (2) the beta to be stuck at 1. This means no change after each training epoch.

Solutions adopted:
- now all sample weights are initialized to 1 and are never normalized
- changed calculation of beta so that sigmoid of error rate isn't saturated (always 1)